# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: N-Pricing API
  description: |
    Bank-grade pricing platform — FTP engine, customer relationships,
    channel quotes, governance dossiers, observability and tenant ops.

    Three families of endpoints:
      1. Pricing Edge Functions (Supabase) — POST /pricing, /pricing/batch,
         /pricing/scenarios. Write pricing_snapshots on every call.
      2. Express server (/api/*) — entity-scoped CRUD + read endpoints behind
         x-entity-id tenancy middleware and JWT auth.
      3. Channel API (/api/channel/*) — API-key authenticated, rate limited
         per key, separate from the JWT path.

    Refer to docs/phase-0-rollout.md for the env vars that gate strict
    tenancy enforcement and mock-fallback rejection.
  version: 2.0.0
  contact:
    name: N-Pricing Team
servers:
  - url: https://{project-ref}.supabase.co/functions/v1
    description: Supabase Edge Functions
  - url: https://{deployment}/api
    description: Express server
security:
  - bearerAuth: []
...
paths:
...
  /notifications/push/vapid-public-key:
    get:
      summary: Return VAPID public key for pushManager.subscribe
      tags: [notifications]
      responses:
        '200': { description: publicKey + configured:true }
        '503': { description: VAPID not configured on this deployment }
  /notifications/push/subscribe:
    post:
      summary: Register a Web Push subscription (UPSERT by entity+user+endpoint)
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint, keysP256dh, keysAuth]
              properties:
                endpoint:    { type: string }
                keysP256dh:  { type: string }
                keysAuth:    { type: string }
                userAgent:   { type: string, nullable: true }
      responses:
        '201': { description: Created or updated }
  /notifications/push/unsubscribe:
    post:
      summary: Remove a Web Push subscription by endpoint
      tags: [notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [endpoint]
              properties:
                endpoint: { type: string }
      responses:
        '200': { description: ok }
...
  /notifications/push/test:
    post:
      summary: Send a test push to all of the user's registered devices
      tags: [notifications]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                message: { type: string }
      responses:
        '200': { description: delivered + total + retried + staleEndpointsPurged + failures }
        '503': { description: VAPID not configured }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...